requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
configparser==3.7.4
cryptography==2.7
idna==2.8
jeepney==0.4
Naked==0.1.31
pycparser==2.19
pycrypto==2.6.1
pykerberos==1.2.1
PyYAML==5.1.1
requests==2.22.0
requests-kerberos==0.12.0
SecretStorage==3.1.1
shellescape==3.4.1
six==1.12.0
soupsieve==1.9.2
urllib3==1.25.3
